requirements.txt
Flask[async]
Flask-Login
Flask-Migrate
SQLAlchemy
pydantic
alembic
gunicorn
python-dotenv
APScheduler
python-telegram-bot
redis
rq
fastapi
uvicorn
httpx
python-multipart
gevent
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
